Able to compile and run application. DLL and .h file problems have been fixed. Next need to test REQUET_AIRPOT retunrs and figure out dwdata pointer problems.

+++ SitAware/Connection.h
@@ -5,7 +5,7 @@
 #include <tuple>
 #include "Questions.h"
 #include "Interrogator.h"
-
+#include "SimConnect.h"
 //Class Connection
 class Connection
 {
@@ -17,7 +17,7 @@ public:
 	static HANDLE hSimConnect;
 
 	static double fsx_result;
-	std::tuple<double, float> answer_buffer;
+	std::tuple<long, long> answer_buffer;
 	double fsx_answer_buffer;
 	double result_buffer;
